pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-finance-tracker"
version = "0.1.0"
description = "Desktop app to track personal income and expenses (Python + PostgreSQL)."
readme = "README.md"
requires-python = ">=3.12"

authors = [
    { name = "Marcos Vinicius Thibes Kemer" }
]

dependencies = [
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0"	
]

[project.gui-scripts]
finance-tracker = "app.gui:main"

[project.scripts]
finance-tracker-checkdb = "app.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
